M5': Slim validate_transition_probs to state-probs-only, move to user_regime.py
Drop the regime-probs overload — that mode is redundant with
`validate_regime_transitions_all_periods` which runs unconditionally
during `model.solve()` and `model.simulate()` and additionally
checks inactive-regime probability and reachability.

Keep the state-probs mode. It is the only defence against four
silent-correctness bug classes in user-written MarkovTransition
functions for states: wrong-shape broadcasting, values outside
[0, 1], rows not summing to 1, and subscript-order swaps relative
to the function signature.

Move slimmed function + helpers (`_extract_markov_transition`,
`_build_grids`, `_build_expected_shape`) to `user_regime.py` next to
the `Regime` class they operate on. Update `lcm/__init__.py` import.
Drop three regime-probs tests from `tests/test_pandas_utils.py`.
Update `docs/user_guide/pandas_interop.md` accordingly. Part of
Phase 1 — see `Phase 1 — Validation Cleanup.md`.
