+++ src/lcm/__init__.py
@@ -72,9 +72,12 @@ from lcm.persistence import (  # noqa: E402
     _bind_forward_refs as _bind_persistence_forward_refs,
 )
 from lcm.simulation.result import SimulationResult  # noqa: E402
-from lcm.user_regime import MarkovTransition, Regime  # noqa: E402
+from lcm.user_regime import (  # noqa: E402
+    MarkovTransition,
+    Regime,
+    validate_transition_probs,
+)
 from lcm.utils.containers import invert_regime_ids  # noqa: E402
-from lcm.utils.error_handling import validate_transition_probs  # noqa: E402
 from lcm.variables import (  # noqa: E402
     _bind_forward_refs as _bind_variables_forward_refs,
 )
